BUG: fix admin schedule loader and home link
The admin schedule run button turns into a spinner when we click it.
But it wasn't disabled. So it was possible to click it and re-run
while it's running. Now the button is disabled while it's running and
spinning.

The admin page has a home: option. I've set it up so that when you're
in an admin sub-page (e.g. users, schedule, webshell), clicking on
the admin icon / title takes you to the admin home page. From there,
if you click on the same admin icon / title again, it takes you to
the application home page specified in the home: option