ENH: better checkbox, radio options - forms
- support multiple checkbox, radio options
- updated field draggable icon while creating a form

checkbox and radio fields could've multiple options. Multiple value submissions must be supported for checkboxes.

While viewing the results (responses tab), the response needs to be sent in the right format.
In the backend, literal_eval is used to accept json field response. The resultant object is concatenated with the default dataframe.
Rows are then collapsed into one to ignore the NaNs in other columns.