FIX: Pipelines support iter=True
build_transform(iter=True) compiles strings into functions that return
iterables. But iter=True is not supported for arrays of strings.

This creates a problem in FunctionHandler, which requires iterables.
(This is to support generators that can yield values one-by-one.)

Now, build_transform() now supports iter=True for an array of strings.
It just returns the output as an array (or as-is, if it's a generator.)