GRAMEX-195 ⁃ FIX: Pipelines support iter=True (#538)
build_transform() compiles a string into a function.
build_transform(iter=True) forces the function to return an array.

Last week, we modified build_transform() to compile pipelines.
(Pipelines = ARRAYS of functions.)
But it didn't support iter=True.
It always returned the value as-is.
(Why? I didn't think it would be needed.)

But there's a reason why iter=True is there in build_transform.
FunctionHandler requires iterables.
This lets it render functions that yield values one by one, slowly.
So, build_transform() needs to support iter=True even for
ARRAYS of functions, i.e. pipelines.

Now, build_transform() now supports iter=True for pipelines.
It just returns the output as an array (or as-is, if it's a generator.)


This apart, there's another change.
Earlier, pipeline steps had to be dicts with a `function:` key, like this:

```yaml
function:
    - {name: x, function: random.randint(0, 100)}
    - {name: y, function: random.randint(0, 100)}
    - {function: x + y}
```

Note the last line. `{function: x + y}`.
Ideally, we should be able to write it like `x + y`.

This commit allows that.