FIX: storelocations.otp purge now works
storelocations.otp needs to be purged of expired keys.
This was handled by a daily schedule.
But this schedule uses time.time() to get the current time.
That requires the time module to be imported first.
(Gramex transforms auto-import only modules at the start.)

So, refactor the purge code into a function.