FIX: OpenAPIHandler spec for FunctionHandler supports POST-only
If a FunctionHandler has only a POST method, the code auto-adds a GET
with a summary. (Because we do that by default for all handlers.)

This commit fixes that.

We were also updating the DEFAULT OpenAPI specs with each handler's specs.
So the specs were over-writing each other.

Now we make a copy of the spec before updating it.