FIX: migration works on UNIX. Fixes #677
Gramex 1.87.0 requires migrating storelocation.user
(auth.user.db) schema. We used BLOB earlier for SQLiteStore.
We're moving to FormHandler, which requires TEXT.

Gramex 1.87.0 migrates the SQLite database to TEXT.

But the code was tested on Windows, not UNIX. On UNIX,
the migration fails because the path comparison uses a
different separator.

This commit fixes the issue.