FIX: mailers use From: as sender if provided. Fixes #752
Earlier, even if mailer.mail(from=) was provided, it would not
override the email service default email: config as the sender.

But this creates a problem with SES, which requires the IAM as
the email: in the email service.

This commit allows the from= to override the email sender.

NOTE: This means that we cannot specify a different From:
than sender.