FIX: render user info only after generating it
We try to render the user object as YAML.
But the user object may have non-YAML Python objects.
This raises an Exception in yaml.safe_dump.

Earlier, we wrote the <pre> and then tried to safe_dump.
If there's an Exception, the <pre> is written, but NOT
the </pre>. That messes up the HTML.

Now, we generate the user info text first and THEN
render the <pre>user_info</pre>